cpu_input_boost: Disable cores 1-3,5-7 on suspend
We can save some power by disabling most of the cores except one from
the little cluster and one from the big cluster. If we disabled all of
the big cluster, the scheduler would have ot migrate all the tasks from
the big cluster to the little cluster, causing havoc.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>